requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.1
certifi==2023.11.17
click==8.1.7
coverage==7.3.3
ecdsa==0.18.0
elastic-transport==8.11.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
exceptiongroup==1.2.0
Faker==20.1.0
fastapi==0.105.0
fastapi-debug-toolbar==0.5.0
flake8==6.1.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
psycopg2==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyflakes==3.1.0
pyinstrument==4.6.1
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
zipp==3.17.0
